+++ docs/source/api-reference/production_configuration.md
@@ -68,6 +68,20 @@ related production-configuration helpers used by the supported workflows.
    :members:
 ```
 
+## production_comparison
+
+```{eval-rst}
+.. automodule:: production_configuration.production_comparison
+   :members:
+```
+
+## production_metadata
+
+```{eval-rst}
+.. automodule:: production_configuration.production_metadata
+   :members:
+```
+
 
 (angle-ranges)=
