+++ src/simtools/production_configuration/production_metadata.py
@@ -1,0 +1,137 @@
+"""Write and validate metadata manifests for completed simulation production jobs."""
+
+from pathlib import Path
+from types import SimpleNamespace
+
+from simtools.io.ascii_handler import write_data_to_file
+from simtools.model.model_utils import read_overwrite_model_parameter_dict
+from simtools.production_configuration.job_grid_io import (
+    job_grid_row_to_simulate_prod_args,
+    read_job_grid,
+)
+from simtools.production_configuration.job_metadata import (
+    REQUIRED_SIMULATION_JOB_METADATA_ARGUMENTS,
+    build_production_job_manifest,
+    build_simulation_job_metadata,
+)
+from simtools.production_configuration.production_file_selection import (
+    SIMULATE_PROD_JOB_METADATA,
+    ProductionManifest,
+    check_manifest,
+    inventory_production_files,
+    validate_required_production_outputs,
+)
+
+
+def write_production_metadata(args_dict):
+    """Write or validate production job metadata manifests.
+
+    Parameters
+    ----------
+    args_dict : dict
+        Application arguments containing ``production_path`` and, in write mode, ``job_grid_file``.
+    """
+    production_path = Path(args_dict["production_path"])
+    if args_dict.get("check"):
+        _check_existing_manifests(production_path)
+        return
+    _write_manifests(production_path, args_dict["job_grid_file"], args_dict)
+
+
+def _check_existing_manifests(production_path):
+    """Check all existing simulate_prod job manifests below a production path."""
+    job_directories = _job_directories(production_path)
+    if not job_directories:
+        raise FileNotFoundError(f"No production job directories found in {production_path}.")
+    missing = [
+        directory / SIMULATE_PROD_JOB_METADATA
+        for directory in job_directories
+        if not (directory / SIMULATE_PROD_JOB_METADATA).is_file()
+    ]
+    if missing:
+        raise FileNotFoundError(
+            "Missing production metadata manifest(s): " + ", ".join(map(str, missing))
+        )
+    for job_directory in job_directories:
+        check_manifest(job_directory / SIMULATE_PROD_JOB_METADATA)
+
+
+def _write_manifests(production_path, job_grid_file, args_dict):
+    """Write manifests for job directories described by a job grid."""
+    rows, metadata = read_job_grid(job_grid_file)
+    for index, row in enumerate(rows):
+        job_directory = production_path / f"job-{index + 1:06d}"
+        if not job_directory.is_dir():
+            raise FileNotFoundError(f"Production job directory not found: {job_directory}")
+        manifest_path = job_directory / SIMULATE_PROD_JOB_METADATA
+        if manifest_path.exists() and not args_dict.get("overwrite"):
+            raise FileExistsError(
+                f"Metadata manifest already exists: {manifest_path}. Use --overwrite to replace it."
+            )
+        resolved_args = job_grid_row_to_simulate_prod_args(row, metadata)
+        resolved_args.setdefault("simulation_software", "corsika_sim_telarray")
+        resolved_args.setdefault("eslope", -2.0)
+        _validate_resolved_configuration(resolved_args, job_grid_file)
+        overwrite_parameter_file = resolved_args.get("overwrite_model_parameters")
+        overwrite_parameters = (
+            read_overwrite_model_parameter_dict(overwrite_parameter_file)
+            if overwrite_parameter_file
+            else {}
+        )
+        array_model = SimpleNamespace(
+            model_version=resolved_args["model_version"],
+            overwrite_model_parameter_dict=overwrite_parameters,
+            array_elements={},
+            site_model=SimpleNamespace(parameters={}),
+        )
+        simulator = SimpleNamespace(
+            run_number=resolved_args["run_number"],
+            array_models=[array_model],
+            corsika_configurations=[],
+        )
+        file_inventory = inventory_production_files(job_directory)
+        validate_required_production_outputs(
+            file_inventory,
+            resolved_args["simulation_software"],
+            job_directory,
+        )
+        manifest = build_production_job_manifest(
+            resolved_args,
+            simulator,
+            job_directory,
+            file_inventory=file_inventory,
+            catalog_metadata=build_simulation_job_metadata(
+                resolved_args, simulator, include_sct=False
+            ),
+            atmosphere_configuration=_backfilled_atmosphere_configuration(resolved_args),
+        )
+        check_manifest(ProductionManifest(path=manifest_path, data=manifest))
+        write_data_to_file(manifest, manifest_path)
+
+
+def _backfilled_atmosphere_configuration(args_dict):
+    """Return only atmosphere values known from the authoritative job grid."""
+    threshold = args_dict.get("curved_atmosphere_min_zenith_angle")
+    return {"curved_atmosphere_min_zenith_angle": threshold} if threshold is not None else {}
+
+
+def _validate_resolved_configuration(args_dict, job_grid_file):
+    """Require fields needed to write truthful production metadata."""
+    missing = [
+        key
+        for key in ("run_number", *REQUIRED_SIMULATION_JOB_METADATA_ARGUMENTS)
+        if args_dict.get(key) is None
+    ]
+    if missing:
+        raise ValueError(
+            f"Job grid {job_grid_file} does not provide required resolved configuration "
+            "field(s): " + ", ".join(missing)
+        )
+
+
+def _job_directories(production_path):
+    """Return sorted direct production job directories."""
+    production_path = Path(production_path)
+    if not production_path.is_dir():
+        raise FileNotFoundError(f"Production path not found: {production_path}")
+    return sorted(path for path in production_path.glob("job-*") if path.is_dir())
